Fix phpstan errors from larastan v3 stricter defaults
- HealthChecker::checkLargeTables: cast score decrement to int — the
  30/count() expression is float|int but the parameter is by-ref int.
- SqlFileLoader::executeSqlFile: switch DB::unprepared() to DB::statement()
  since the loaded SQL is a string, not a literal-string. statement() runs
  the same prepared path and works for DDL.
- TestCase::getEnvironmentSetUp: read env via $_ENV superglobal instead of
  Laravel's env() helper (noEnvCallsOutsideOfConfig is now a default).
- HealthCheckTest, PgHelperLibraryTest, SelectiveOperationsTest: drop
  trivially-true assertIsString/assertIsInt/assertIsFloat assertions now
  that the array shapes are statically typed.
- PgHelperLibraryTest: remove stale @phpstan-ignore-line comment from an
  assertion that no longer reports an error.
- tests/Unit/PlaceholderTest.php: removed — the file's TODO asked to
  delete it once real tests existed; Validators/ has 9 real unit tests.

+++ src/Libraries/Helpers/HealthChecker.php
@@ -203,7 +203,7 @@ class HealthChecker
         ");
 
         foreach ($largeTables as $table) {
-            $score -= min(10, 30 / max(1, \count($largeTables)));
+            $score -= (int) min(10, 30 / max(1, \count($largeTables)));
             $issues[] = "Large table: {$table->tablename} ({$table->size})";
         }
     }
